% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quote_data.R
\name{get_quote_data}
\alias{get_quote_data}
\title{get_quote_data}
\usage{
get_quote_data(character = "", movie = "")
}
\arguments{
\item{character}{A string indicating what character quotes should be filtered by. \code{get_quote_data} employs string matching, so for example entering "Frodo" will specify "Frodo baggins" \emph{and} "Frodo Gardner".}

\item{movie}{A string indicating what movie quotes should be filtered by. \code{get_quote_data} employs string matching, so for example entering "Frodo" will specify "Frodo baggins" \emph{and}}
}
\value{
A dataframe of quotes made by a set of characters in a set of lotr films.
}
\description{
\code{get_quote_data} is a function that uses \code{lotr}, another function in this package, and formats the data.
}
