% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotr.R
\name{lotr}
\alias{lotr}
\title{lotr()}
\source{
\href{https://the-one-api.dev/}{The one API}
}
\usage{
lotr(end_point_argument = "movie")
}
\arguments{
\item{end_point_argument}{A string (one of "character", "movie", "book", "quotes" etc) that specifies the end point of the lotr API that should be called in the GET request.}
}
\value{
A dataframe of the data that was returned from the API.
}
\description{
A function for connection with the LOTR API. There is a rate limit: 100 requests per 10 min, or one every 6 seconds.
}
