% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotr.R
\name{lotr}
\alias{lotr}
\title{LOTR}
\source{
\href{https://the-one-api.dev/}{The one API}
}
\usage{
lotr(arg1 = "character", arg2)
}
\arguments{
\item{arg1}{A short description of the arg1 argument. Must specify: accepted format.}

\item{arg2}{A short description of the arg2 argument. Must specify: accepted format.}
}
\value{
A short description of the output format.
}
\description{
A function for connection with the LOTR API.
}
